chore(crdt-charts): refactor into separate files
